jwt authentication

+++ routes/tasks.js
@@ -2,20 +2,18 @@ module.exports = app => {
     const Tasks = app.db.models.Tasks;
 
     app.route("/tasks")
-        // .all((req, res, next) => {
-        //     //Middleware de pré-execução das rotas
-        //     delete req.body.id;
-        //     next();
-        // })
+        .all(app.auth.authenticate())
         .get((req, res) => {
             // lista de tarefas
-            Tasks.findAll({})
+            Tasks.findAll({where: {user_id: req.user.id}})
                 .then(result => res.json(result))
                 .catch(error => {
                     res.status(412).json({msg: error.message});
                 });
         })
         .post((req, res) => {
+            //authentica
+            req.body.user_id = req.user.id;
             //cadastra nova tarefa
             Tasks.create(req.body)
             .then(result => res.json(result))
@@ -25,14 +23,10 @@ module.exports = app => {
         });
 
     app.route("/tasks/:id")
-        // .all((req, res, next) => {
-        //     //Middleware de pré-execução das rotas
-        //     delete req.body.id;
-        //     next();
-        // })
-        .get((req, res) => {
+    .all(app.auth.authenticate())
+    .get((req, res) => {
             // "/tasks/1" lista tarefa
-            Tasks.findOne({where: req.params})
+            Tasks.findOne({where: {id: req.params.id, user_id: req.user.id}})
                 .then(result => {
                     if(result) {
                         res.json(result);
@@ -47,7 +41,7 @@ module.exports = app => {
         })
         .put((req, res) => {
             // "/tasks/1" cadastra nova tarefa
-            Tasks.update(req.body, {where: req.params})
+            Tasks.update(req.body, {where: {id: req.params.id, user_id: req.user.id}})
             .then(result => res.sendStatus(204))
             .catch(error => {
                 res.status(412).json({msg: error.message});
@@ -55,7 +49,7 @@ module.exports = app => {
         })
         .delete((req, res) => {
             // "/tasks/1" exclui tarefa
-            Tasks.destroy({where: req.params})
+            Tasks.destroy({where: {id: req.params.id, user_id: req.user.id}})
                 .then(result => res.sendStatus(204))
                 .catch(error => {
                     res.status(412).json({msg: error.message});
